// src/types/tvCreditsInterface.ts
// Generated by https://quicktype.io

export interface TvCredits {
  cast: Cast[];
  crew: any[];
  id: number;
}

export interface Cast {
  adult: boolean;
  gender: number;
  id: number;
  known_for_department: KnownForDepartment;
  name: string;
  original_name: string;
  popularity: number;
  profile_path: null | string;
  character: string;
  credit_id: string;
  order: number;
}

export enum KnownForDepartment {
  Acting = 'Acting',
}