// src/types/multiSearch.ts
// Generated by https://quicktype.io

export interface MultiSearch {
  page: number;
  results: SearchResult[];
  total_pages: number;
  total_results: number;
}

export interface SearchResult {
  adult?: boolean;
  backdrop_path?: null | string;
  genre_ids?: number[];
  id: number;
  media_type: MediaType;
  original_language?: OriginalLanguage;
  original_title?: string;
  overview?: string;
  popularity: number;
  poster_path?: null | string;
  release_date?: string;
  title?: string;
  video?: boolean;
  vote_average?: number;
  vote_count?: number;
  gender?: number;
  known_for?: KnownFor[];
  known_for_department?: null | string;
  name?: Name;
  profile_path?: null | string;
  first_air_date?: string;
  origin_country?: any[];
  original_name?: Name;
}

export interface KnownFor {
  adult: boolean;
  backdrop_path: string;
  genre_ids: number[];
  id: number;
  media_type: MediaType;
  original_language: OriginalLanguage;
  original_title: string;
  overview: string;
  poster_path: string;
  release_date: string;
  title: string;
  video: boolean;
  vote_average: number;
  vote_count: number;
}

export enum MediaType {
  Movie = 'movie',
  Person = 'person',
  Tv = 'tv',
}

export enum OriginalLanguage {
  En = 'en',
}

export enum Name {
  AllenActor = 'Allen Actor',
  ExtremeActor = 'Extreme Actor',
  RepeatedActor = 'Repeated Actor',
}